ADD: id validation

+++ src/controller/department.controller.ts
@@ -6,6 +6,8 @@ import DepartmentService from '../service/department.service'
 import DepartmentDto from '../dto/department.dto'
 import validator from '../middleware/validator.middleware'
 import EditDepartmentDto from '../dto/edit-department.dto'
+import BadRequestException from '../exception/bad-request.exception'
+import { isUUID } from 'class-validator'
 
 class DepartmentController {
   public router: Router
@@ -65,6 +67,8 @@ class DepartmentController {
   ) => {
     try {
       const departmentId = req.params.id
+      if (!isUUID(departmentId))
+        throw new BadRequestException('Invalid department id')
 
       const department = await this.departmentService.getDepartmentById(
         departmentId
@@ -85,6 +89,8 @@ class DepartmentController {
   ) => {
     try {
       const departmentId = req.params.id
+      if (!isUUID(departmentId))
+        throw new BadRequestException('Invalid department id')
 
       const department = await this.departmentService.editDepartment(
         departmentId,
@@ -125,6 +131,8 @@ class DepartmentController {
   ) => {
     try {
       const departmentId = req.params.id
+      if (!isUUID(departmentId))
+        throw new BadRequestException('Invalid department id')
 
       const updatedDepartment = await this.departmentService.updateDepartment(
         departmentId,
@@ -147,6 +155,8 @@ class DepartmentController {
   ) => {
     try {
       const departmentId = req.params.id
+      if (!isUUID(departmentId))
+        throw new BadRequestException('Invalid department id')
 
       await this.departmentService.removeDepartmentById(departmentId)
       res.status(204)
