+++ src/controller/shelf.controller.ts
@@ -3,7 +3,8 @@ import ShelfService from '../service/shelf.service'
 import ShelfDto from '../dto/shelf.dto'
 import validator from '../middleware/validator.middleware'
 import EditShelfDto from '../dto/edit-shelf.dto'
-
+import { isUUID } from 'class-validator'
+import BadRequestException from '../exception/bad-request.exception'
 
 class ShelfController {
   public router: Router
@@ -39,6 +40,7 @@ class ShelfController {
   ) => {
     try {
       const shelfId = req.params.id
+      if (!isUUID(shelfId)) throw new BadRequestException('Invalid shelf id')
 
       const shelf = await this.shelfService.getShelfById(shelfId)
       res.status(200)
@@ -72,6 +74,7 @@ class ShelfController {
   ) => {
     try {
       const shelfId = req.params.id
+      if (!isUUID(shelfId)) throw new BadRequestException('Invalid shelf id')
 
       const editedShelf = await this.shelfService.editShelf(shelfId, req.body)
       res.status(200)
@@ -89,6 +92,7 @@ class ShelfController {
   ) => {
     try {
       const shelfId = req.params.id
+      if (!isUUID(shelfId)) throw new BadRequestException('Invalid shelf id')
 
       const updatedShelf = await this.shelfService.updateShelf(
         shelfId,
@@ -109,6 +113,7 @@ class ShelfController {
   ) => {
     try {
       const shelfId = req.params.id
+      if (!isUUID(shelfId)) throw new BadRequestException('Invalid shelf id')
 
       const editedShelf = await this.shelfService.removeShelf(shelfId)
       res.status(204)
